lepton_module: New vsync handler
VSYNC doesn't start SPI DMA transfers yet, but a quick test on a BBB
with a SPI clock jumpered to the VSYNC input (P8 pin 17) reported many
VSYNC events after running 'dd if=/dev/zero of=/dev/spidev2.0' (need to
ctrl-c out of it).
Checked in new .dts file that adds the VSYNC GPIO and configures the SPI
bus pins.